update octokit for gihub stuff make heals use distance3d instead of distance cache distance calls

+++ E3NextUI/Update.cs
@@ -9,7 +9,7 @@ namespace E3NextUI
     public partial class Update : Form
     {
         public GitHubClient client;
-        public int latestID;
+        public Int64 latestID;
         public Update()
         {
             InitializeComponent();
